requirements.txt
requests==2.32.5
PyYAML==6.0.3
pandas==2.2.3
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
python-dotenv==1.0.0
prefect>=3.0,<4.0
asyncpg
